get distance from pixel, calculate log end with angle of pixel to camera and distance to begin of trailer

+++ include/CameraHandler.hpp
@@ -9,7 +9,7 @@ public:
     CameraHandler();
 
     sl::Mat GetImage();
-    float GetDistance(int x, int y);
+    double GetDistance(int x, int y);
 private:
 
     sl::Camera zed_;
